roles/container-engine/docker/vars/ubuntu-amd64.yml: set repo keys to empty list
We provide a debian mirror containing only a subset of packages. As the
gpg signature is on the repository index, returning a subset of packages
isn't possible without re-signing.

As we don't have dockers gpg key, we want to make this overridable in
the longer run and upstream these eventually.

For now, this is what's necessary to install docker with a custom
package source.

See https://github.com/kubernetes-sigs/kubespray/issues/6879